Refactored code into different files.
Made it easy to debug and change code for staff profile by refactoring code into different files.
